added the idea of a parallel execution class that states on which task a parallel execution will start and where it will end. this could have been done implicitly but then some more calculation would have been necessary.
changed all lists in entity manager to hashtable for quicker reference as O(1) on hashtable but O(n) on list, so changed this where it was right

introduced the predecessor Tasks for tasks that are not start task. this is needed to allow for parallelExecution as we need to know on which tasks we have to wait

made some better error prints an some system exits when problem in importer occur as the simulation wouldnt work anyway with missing entities

next Step is to make ParallelExecution actually Work and to import a Workflow as this has been done by hardcoding, now this will be done dynamically as the workflow with its start end and innertask will be modelled in the ontology. here also the parallel Execution will be linked. And as there may be at some point multiple Parallel Executions on a Workflow it seemed good practice to introduce this as a own class. similar ways may be helpful when conditions will be introduced