added a resource Manager and an pretty linear way to assert eneded resources to tasks, meaning that there is no prioritizing in between tasks, they are taken by the order they are sitting int the waitingTasks map in the Workflow that is looking for the next TRC to be executed on a free Task
Currently it is only worked with unlimited resources, but with the current implementation only a few steps are needed to make use of limited number of resources, f.e. change the ontology and make some adjustments to ResourceType and Resource class. that is a TODO

Furthermore i deleted the Resource class and all Properties using this class in the ontology as everything will be modelled wihtin the resource type class where you say if it should be unlimited resources of that type or specify the number. The Sim will instantiate a limited number if given and if unlimited just instantiates one on demand. Unlimited resource are not saved or reused and will be collected ba the garbage colelctor after a task finishes and ist ressources are freeeeeeeeeeeeeee